Revert "a series bugfix & optimization"
