+++ ui/zenoedit/nodesys/zenosubgraphscene.cpp
@@ -21,10 +21,6 @@
 #include "acceptor/transferacceptor.h"
 #include <zenoui/comctrl/gv/zenoparamwidget.h>
 
-#include "zenomainwindow.h"
-#include <zenovis/ObjectsManager.h>
-#include <viewportinteraction/picker.h>
-
 
 ZenoSubGraphScene::ZenoSubGraphScene(QObject *parent)
     : QGraphicsScene(parent)
@@ -854,26 +850,6 @@ void ZenoSubGraphScene::updateLinkPos(ZenoNode* pNode, QPointF newPos)
     }
 }
 
-void ZenoSubGraphScene::selectObjViaNodes() {
-    // FIXME temp function for merge
-    // for selecting objects in viewport via selected nodes
-    auto scene = Zenovis::GetInstance().getSession()->get_scene();
-    QList<QGraphicsItem*> selItems = this->selectedItems();
-    auto& picker = zeno::Picker::GetInstance();
-    picker.clear();
-    for (auto item : selItems) {
-        if (auto* pNode = qgraphicsitem_cast<ZenoNode*>(item)) {
-            auto node_id = pNode->index().data(ROLE_OBJID).toString().toStdString();
-            for (const auto& [prim_name, _] : scene->objectsMan->pairsShared()) {
-                if (prim_name.find(node_id) != std::string::npos)
-                    picker.add(prim_name);
-            }
-        }
-    }
-    picker.sync_to_scene();
-    zenoApp->getMainWindow()->updateViewport();
-}
-
 void ZenoSubGraphScene::keyPressEvent(QKeyEvent* event)
 {
     QGraphicsScene::keyPressEvent(event);
@@ -919,9 +895,5 @@ void ZenoSubGraphScene::keyPressEvent(QKeyEvent* event)
                 pGraphsModel->endTransaction();
             }
         }
-    }
-    else if (!event->isAccepted() && (event->modifiers() & Qt::ControlModifier) && event->key() == Qt::Key_G) {
-        // FIXME temp function for merge
-        selectObjViaNodes();
     }
 }
