+++ zenovis/include/zenovis/Camera.h
@@ -25,8 +25,6 @@ struct Camera {
     glm::vec3 m_lodfront{0, 0, 1};
     glm::vec3 m_lodup{0, 1, 0};
 
-    bool m_need_sync = false;
-
     struct ZxxHappyLookParam {
         float cx = 0;
         float cy = 0;
@@ -38,9 +36,7 @@ struct Camera {
         bool ortho_mode = false;
         float aperture = 0;
         float focalPlaneDistance = 0;
-    };
-    struct ZxxHappyLookParam m_zxx;
-    struct ZxxHappyLookParam m_zxx_in;
+    } m_zxx;
 
     float getAspect() const {
         return (float)m_nx / (float)m_ny;
