+++ zenovis/include/zenovis/bate/IGraphic.h
@@ -67,7 +67,7 @@ struct IGraphicHandler : IGraphicDraw {
 struct IPicker : IGraphicDraw {
     virtual std::string getPicked(int x, int y) = 0;
     virtual std::string getPicked(int x0, int y0, int x1, int y1) = 0;
-    virtual void focus(const std::string& prim_name) = 0;
+    virtual void setPrimSet(const std::vector<std::string>& prims) = 0;
 };
 
 struct MakeGraphicVisitor {
