+++ zenovis/src/bate/FrameBufferPicker.cpp
@@ -12,7 +12,6 @@
 #include <unordered_map>
 #include <fstream>
 #include <random>
-#include <algorithm>
 
 namespace zenovis {
 namespace {
@@ -181,7 +180,7 @@ static void load_buffer_to_image(unsigned int* ids, int w, int h, const std::str
 // framebuffer picker referring to https://doc.yonyoucloud.com/doc/wiki/project/modern-opengl-tutorial/tutorial29.html
 struct FrameBufferPicker : IPicker {
     Scene* scene;
-    string focus_prim_name;
+    vector<string> prim_set;
 
     unique_ptr<FBO> fbo;
     unique_ptr<Texture> picking_texture;
@@ -221,31 +220,6 @@ struct FrameBufferPicker : IPicker {
     };
 
     explicit FrameBufferPicker(Scene* s) : scene(s) {
-        // generate buffers
-        generate_buffers();
-
-        // generate draw buffer
-        vbo = make_unique<Buffer>(GL_ARRAY_BUFFER);
-        ebo = make_unique<Buffer>(GL_ELEMENT_ARRAY_BUFFER);
-        vao = make_unique<VAO>();
-
-        // unbind fbo & texture
-        CHECK_GL(glBindTexture(GL_TEXTURE_2D, 0));
-        fbo->unbind();
-
-        // prepare shaders
-        obj_shader = scene->shaderMan->compile_program(obj_vert_code, obj_frag_code);
-        vert_shader = scene->shaderMan->compile_program(vert_vert_code, vert_frag_code);
-        prim_shader = scene->shaderMan->compile_program(obj_vert_code, prim_frag_code);
-        empty_shader = scene->shaderMan->compile_program(obj_vert_code, empty_frag_code);
-        empty_and_offset_shader = scene->shaderMan->compile_program(obj_vert_code, empty_and_offset_frag_code);
-    }
-
-    ~FrameBufferPicker() {
-        destroy_buffers();
-    }
-
-    void generate_buffers() {
         // generate framebuffer
         fbo = make_unique<FBO>();
         CHECK_GL(glBindFramebuffer(GL_DRAW_FRAMEBUFFER, fbo->fbo));
@@ -272,9 +246,25 @@ struct FrameBufferPicker : IPicker {
 
         // check fbo
         if(!fbo->complete()) printf("fbo error\n");
+
+        // generate draw buffer
+        vbo = make_unique<Buffer>(GL_ARRAY_BUFFER);
+        ebo = make_unique<Buffer>(GL_ELEMENT_ARRAY_BUFFER);
+        vao = make_unique<VAO>();
+
+        // unbind fbo & texture
+        CHECK_GL(glBindTexture(GL_TEXTURE_2D, 0));
+        fbo->unbind();
+
+        // prepare shaders
+        obj_shader = scene->shaderMan->compile_program(obj_vert_code, obj_frag_code);
+        vert_shader = scene->shaderMan->compile_program(vert_vert_code, vert_frag_code);
+        prim_shader = scene->shaderMan->compile_program(obj_vert_code, prim_frag_code);
+        empty_shader = scene->shaderMan->compile_program(obj_vert_code, empty_frag_code);
+        empty_and_offset_shader = scene->shaderMan->compile_program(obj_vert_code, empty_and_offset_frag_code);
     }
 
-    void destroy_buffers() {
+    ~FrameBufferPicker() {
         if (fbo->fbo) CHECK_GL(glDeleteFramebuffers(1, &fbo->fbo));
         if (picking_texture->tex) CHECK_GL(glDeleteTextures(1, &picking_texture->tex));
         if (depth_texture->tex) CHECK_GL(glDeleteTextures(1, &depth_texture->tex));
@@ -286,24 +276,36 @@ struct FrameBufferPicker : IPicker {
         CHECK_GL(glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT));
 
         // construct prim set
-        vector<std::pair<string, std::shared_ptr<zeno::IObject>>> prims;
+        vector<std::pair<string, PrimitiveObject*>> prims;
         auto prims_shared = scene->objectsMan->pairsShared();
-        if (!focus_prim_name.empty()) {
-            std::shared_ptr<zeno::IObject> focus_prim;
-            for (const auto& [k, v] : prims_shared) {
-                if (focus_prim_name == k)
-                    focus_prim = v;
+        for (const auto& prim_name : prim_set) {
+            PrimitiveObject* prim = nullptr;
+            auto optional_prim = scene->objectsMan->get(prim_name);
+            if (optional_prim.has_value())
+                prim = dynamic_cast<PrimitiveObject*>(scene->objectsMan->get(prim_name).value());
+            else {
+                auto node_id = prim_name.substr(0, prim_name.find_first_of(':'));
+                for (const auto& [n, p] : scene->objectsMan->pairsShared()) {
+                    if (n.find(node_id) != std::string::npos) {
+                        prim = dynamic_cast<PrimitiveObject*>(p.get());
+                        break;
+                    }
+                }
             }
-            if (focus_prim) prims.emplace_back(focus_prim_name, focus_prim);
+            if (prim) prims.emplace_back(std::make_pair(prim_name, prim));
         }
-        else
-            prims = std::move(prims_shared);
+        if (prims.empty()) {
+            for (const auto& [prim_name, prim] : prims_shared) {
+                auto p = dynamic_cast<PrimitiveObject*>(prim.get());
+                if (p) prims.emplace_back(std::make_pair(prim_name, p));
+            }
+        }
 
         // shading primitive objects
         for (unsigned int id = 0; id < prims.size(); id++) {
             auto it = prims.begin() + id;
-            auto prim = dynamic_cast<PrimitiveObject*>(it->second.get());
-            if (prim && prim->has_attr("pos")) {
+            auto prim = it->second;
+            if (prim->has_attr("pos")) {
                 // prepare vertices data
                 auto const &pos = prim->attr<zeno::vec3f>("pos");
                 auto vertex_count = prim->size();
@@ -434,14 +436,7 @@ struct FrameBufferPicker : IPicker {
     }
 
     virtual string getPicked(int x, int y) override {
-        // re-generate buffers for possible window resize
-        destroy_buffers();
-        generate_buffers();
-
-        // draw framebuffer
         draw();
-
-        // check fbo
         if (!fbo->complete()) return "";
         CHECK_GL(glBindFramebuffer(GL_READ_FRAMEBUFFER, fbo->fbo));
         CHECK_GL(glReadBuffer(GL_COLOR_ATTACHMENT0));
@@ -487,14 +482,7 @@ struct FrameBufferPicker : IPicker {
     }
 
     virtual string getPicked(int x0, int y0, int x1, int y1) override {
-        // re-generate buffers for possible window resize
-        destroy_buffers();
-        generate_buffers();
-
-        // draw framebuffer
         draw();
-
-        // check fbo
         if (!fbo->complete()) return "";
 
         // prepare fbo
@@ -559,8 +547,9 @@ struct FrameBufferPicker : IPicker {
         return result;
     }
 
-    virtual void focus(const std::string& prim_name) override {
-        focus_prim_name = prim_name;
+    virtual void setPrimSet(const std::vector<std::string>& prims) override {
+        prim_set.clear();
+        prim_set.assign(prims.begin(), prims.end());
     }
 };
